+++ R/format.R
@@ -28,6 +28,12 @@ format.ggeffects <- function(x,
   # data frame including attributes, that's why this code comes first
   nrow_to_print <- .nrows_to_print(x, n)
   focal_terms <- attributes(x)$terms
+  ci_level <- attributes(x)$ci.lvl
+  response_name <- attributes(x)$response.name
+  # default name, used later
+  if (is.null(response_name)) {
+    response_name <- "Response Level"
+  }
 
   # fix terms for survival models
   a1 <- attr(x, "fitfun", exact = TRUE)
@@ -77,13 +83,18 @@ format.ggeffects <- function(x,
   # check which columns we have - we want to sort by "subgroups"
   sort_columns <- c("response.level", "group", "facet", "panel")[c(has_response, has_groups, has_facets, has_panel)]
 
+  # response needs to be factor
+  if (has_response) {
+    x$response.level <- factor(x$response.level, levels = unique(x$response.level))
+  }
+
   if (length(sort_columns)) {
     insight::check_if_installed("datawizard")
     x <- datawizard::data_arrange(x, sort_columns)
   }
 
   # format column names, to make it work with insight-formatting-functions
-  x$CI <- 0.95
+  x$CI <- ifelse(is.null(ci_level), 95, ci_level)
   colnames(x)[colnames(x) == "conf.low"] <- "CI_low"
   colnames(x)[colnames(x) == "conf.high"] <- "CI_high"
 
@@ -113,7 +124,7 @@ format.ggeffects <- function(x,
       for (i in sort_columns) {
         prefix <- switch(
           i,
-          response.level = "Response level",
+          response.level = response_name,
           group = ifelse(length(focal_terms) > 1, focal_terms[2], ""),
           facet = ifelse(length(focal_terms) > 2, focal_terms[3], ""),
           panel = ifelse(length(focal_terms) > 3, focal_terms[4], ""),
@@ -155,12 +166,18 @@ format.ggeffects <- function(x,
     if (isTRUE(collapse_tables)) {
       insight::check_if_installed("datawizard")
       # first focal term is main term, we don't want to touch it here
-      x <- datawizard::data_rename(x, sort_columns, focal_terms[-1])
-      x <- datawizard::data_relocate(x, focal_terms[-1], after = 1)
+      focal_terms <- focal_terms[-1]
+      # if we have ordinal models and alike, we have a response level, which
+      # we need to include as first value of "focal_terms"
+      if (has_response) {
+        focal_terms <- c(response_name, focal_terms)
+      }
+      x <- datawizard::data_rename(x, sort_columns, focal_terms)
+      x <- datawizard::data_relocate(x, focal_terms, after = 1)
       # we need to remove "groups", else table will be separated again
       x$groups <- NULL
       # remove repeating elements in focal term columns
-      for (i in focal_terms[-1]) {
+      for (i in focal_terms) {
         for (j in nrow(x):2) {
           if (x[[i]][j] == x[[i]][j - 1]) x[[i]][j] <- ""
         }