+++ R/utils_ggpredict.R
@@ -1,3 +1,79 @@
+.validate_type_argument <- function(model, type, ppd, marginaleffects = FALSE) {
+  # marginaleffects supports the predict-method types
+  # we need a different approach to validation here
+  if (marginaleffects) {
+    # first, we overwrite the "default"
+    if (type == "fixed") {
+      if (class(model)[1] %in% .default_type$class) {
+        type <- .default_type$type[.default_type$class == class(model)[1]]
+      } else {
+        type <- "response"
+      }
+    }
+    # check which types are supported by the model's predict-method
+    type_options <- .typedic$type[.typedic$class == class(model)[1]]
+    if (!type %in% c("response", type_options)) {
+      insight::format_error(sprintf(
+        "`type = \"%s\"` is not supported. Please use %s%s.",
+        type,
+        if (length(type_options) > 1) "one of " else "",
+        toString(paste0("`", type_options, "`"))
+      ))
+    }
+    return(list(type = type, ppd = ppd))
+  }
+
+  # if we call "predict()" or "emmeans()", we have these different options
+  type <- match.arg(type, choices = c(
+    "fe", "fixed", "count", "re", "random",
+    "fe.zi", "zero_inflated", "re.zi", "zi_random",
+    "zero_inflated_random", "zi.prob", "zi_prob",
+    "sim", "simulate", "surv", "survival", "cumhaz",
+    "cumulative_hazard", "sim_re", "simulate_random",
+    "debug", "fixed_ppd", "random_ppd"
+  ))
+  # handle Bayes exceptions for type with ppd
+  if (type %in% c("fixed_ppd", "random_ppd")) {
+    ppd <- TRUE
+    type <- gsub("_ppd", "", type, fixed = TRUE)
+  }
+
+  type <- switch(type,
+    fixed = ,
+    count = "fe",
+    random = "re",
+    zi = ,
+    zero_inflated = "fe.zi",
+    zi_random = ,
+    zero_inflated_random = "re.zi",
+    zi_prob = "zi.prob",
+    survival = "surv",
+    cumulative_hazard = "cumhaz",
+    simulate = "sim",
+    simulate_random = "sim_re",
+    type
+  )
+
+  list(type = type, ppd = ppd)
+}
+
+
+.retrieve_type_option <- function(model) {
+  # retrieve model object's predict-method prediction-types (if any)
+  predict_method <- .safe(lapply(
+    class(model), function(i) {
+      utils::getS3method("predict", i)
+    }
+  ))
+  # check whether model class has a predict method
+  if (!is.null(predict_method)) {
+    predict_method <- predict_method[!vapply(predict_method, is.null, TRUE)][[1]]
+  }
+  # retrieve model object's predict-method prediction-types (if any)
+  .safe(suppressWarnings(eval(formals(predict_method)$type)))
+}
+
+
 .back_transform_response <- function(model, mydf, back.transform, response.name = NULL, verbose = TRUE) {
   # skip if no information available
   if (is.null(model) && is.null(response.name)) {