+++ tests/testthat/test-glm.R
@@ -131,3 +131,20 @@ withr::with_environment(
     expect_s3_class(ggemmeans(m4, "period"), "data.frame")
   })
 )
+
+
+test_that("ggaverage, invlink", {
+  skip_if_not_installed("marginaleffects")
+  dat2 <- data.frame(
+    sex = c("m", "w", "w", "m", "w", "w", "m", "w", "w", "m", "m", "w", "w"),
+    smoking = c(0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 0, 1),
+    age = c(10, 45, 50, 40, 45, 12, 14, 55, 60, 10, 14, 50, 40),
+    stringsAsFactors = FALSE
+  )
+  m3 <- glm(smoking ~ sex, data = dat2, family = binomial("logit"))
+  out1 <- ggaverage(m3, "sex")
+  out2 <- marginaleffects::avg_predictions(m3, variables = "sex", type = "invlink(link)")
+  expect_equal(out1$predicted, out2$estimate, tolerance = 1e-4)
+  expect_equal(out1$conf.low, out2$conf.low, tolerance = 1e-4)
+  expect_equal(out1$conf.high, out2$conf.high, tolerance = 1e-4)
+})